@@ -18,54 +18,54 @@
   },
   "private": true,
   "dependencies": {
-    "@agm/core": "^1.0.0-beta.5",
-    "@angular/animations": "^7.2.15",
-    "@angular/common": "^7.2.15",
-    "@angular/compiler": "^7.2.15",
-    "@angular/core": "^7.2.15",
-    "@angular/fire": "^5.1.1",
-    "@angular/forms": "^7.2.15",
-    "@angular/platform-browser": "^7.2.15",
-    "@angular/platform-browser-dynamic": "^7.2.15",
-    "@angular/router": "^7.2.15",
-    "@fortawesome/fontawesome-free": "^5.9.0",
+    "@agm/core": "^1.0.0",
+    "@angular/animations": "^8.2.7",
+    "@angular/common": "^8.2.7",
+    "@angular/compiler": "^8.2.7",
+    "@angular/core": "^8.2.7",
+    "@angular/fire": "^5.2.1",
+    "@angular/forms": "^8.2.7",
+    "@angular/platform-browser": "^8.2.7",
+    "@angular/platform-browser-dynamic": "^8.2.7",
+    "@angular/router": "^8.2.7",
+    "@fortawesome/fontawesome-free": "^5.10.1",
     "@ngx-translate/core": "^11.0.1",
     "@ngx-translate/http-loader": "^4.0.0",
-    "angular-google-charts": "^0.1.3",
+    "angular-google-charts": "^0.1.6",
     "angulartics2": "^7.5.2",
-    "core-js": "^2.6.8",
-    "firebase": "^5.8.6",
-    "materialize-css": "~1.0.0",
-    "rxjs": "~6.5.2",
-    "ts-md5": "^1.2.4",
-    "tslib": "^1.9.0",
-    "zone.js": "~0.8.26"
+    "core-js": "^2.6.9",
+    "firebase": "^6.3.4",
+    "materialize-css": "~1.0.0-rc.2",
+    "rxjs": "^6.5.3",
+    "ts-md5": "^1.2.6",
+    "tslib": "^1.10.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.9",
-    "@angular/cli": "^7.3.9",
-    "@angular/compiler-cli": "^7.2.15",
-    "@angular/language-service": "^7.2.15",
+    "@angular-devkit/build-angular": "~0.803.5",
+    "@angular/cli": "^8.3.5",
+    "@angular/compiler-cli": "^8.2.7",
+    "@angular/language-service": "^8.2.7",
     "@biesbjerg/ngx-translate-extract": "^2.3.4",
-    "@types/jasmine": "~2.8.8",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/materialize-css": "~1.0.6",
-    "@types/node": "~8.9.4",
-    "codelyzer": "~4.5.0",
-    "express": "^4.16.4",
-    "jasmine-core": "~2.99.1",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^4.0.1",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
+    "@types/jasmine": "^2.8.16",
+    "@types/jasminewd2": "^2.0.6",
+    "@types/materialize-css": "^1.0.6",
+    "@types/node": "^8.9.5",
+    "codelyzer": "^5.0.1",
+    "express": "^4.17.1",
+    "jasmine-core": "^2.99.1",
+    "jasmine-spec-reporter": "^4.2.1",
+    "karma": "^4.3.0",
+    "karma-chrome-launcher": "^2.2.0",
+    "karma-coverage-istanbul-reporter": "^2.0.6",
+    "karma-jasmine": "^1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
     "mz": "^2.7.0",
     "npm-run-all": "^4.1.5",
-    "protractor": "~5.4.0",
-    "puppeteer": "^1.19.0",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "~3.2.2"
+    "protractor": "^5.4.2",
+    "puppeteer": "^1.20.0",
+    "ts-node": "^7.0.1",
+    "tslint": "^5.11.0",
+    "typescript": "~3.5.3"
   }
 }
